# swagger/swagger.json
{"info":{"title":"ZeroSlope","version":"1.0.0","description":"ZeroSlope API developed in Node."},"swagger":"2.0","paths":{"/auth/authenticate":{"post":{"tags":["Authentication"],"summary":"Generates a JWT token","produces":["application/json"],"parameters":[{"name":"Model","description":"Token Request","in":"body","required":true,"schema":{"$ref":"#/definitions/TokenRequest"}}],"responses":{"200":{"description":"Returns a valid JWT token"}}}},"/health":{"get":{"tags":["Health"],"summary":"Determines if the service is in a healthy state","produces":["application/json"],"responses":{"200":{"description":"The service is in a healthy state"}}}},"/sample":{"get":{"tags":["Sample"],"summary":"Gets a list of sample models","description":"Returns a list of sample models","produces":["application/json"],"responses":{"200":{"description":"An array of sample models","schema":{"$ref":"#/definitions/SampleResponse"}}}},"post":{"tags":["Sample"],"summary":"Creates and stores a sample model","description":"Returns a list of sample models","produces":["application/json"],"parameters":[{"name":"Model","description":"Sample object","in":"body","required":true,"schema":{"$ref":"#/definitions/SampleRequest"}}],"responses":{"200":{"description":"Successfully created"}}}},"/sample/{id}":{"get":{"tags":["Sample"],"summary":"Reads a single sample model by id","description":"Returns a single of sample model","produces":["application/json"],"parameters":[{"name":"id","description":"Sample model's id","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"An single sample model","schema":{"$ref":"#/definitions/SampleResponse"}}}},"put":{"tags":["Sample"],"summary":"Updates and stores a sample model","description":"Updates and stores a sample model","produces":"application/json","parameters":[{"name":"id","description":"Sample model's id","in":"path","required":true,"type":"string"},{"name":"Model","in":"body","description":"Fields for the Sample resource","schema":{"type":"array","$ref":"#/definitions/SampleRequest"}}],"responses":{"200":{"description":"Successfully updated"}}},"delete":{"tags":["Sample"],"summary":"Deletes a sample mode by id","description":"Deletes a single sample by id","produces":["application/json"],"parameters":[{"name":"id","description":"Samples's id","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Successfully deleted"}}}}},"definitions":{"TokenRequest":{"properties":{"client_id":{"type":"string"},"secret":{"type":"string"}}},"SampleResponse":{"properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"}}},"SampleRequest":{"properties":{"name":{"type":"string"},"description":{"type":"string"}}}},"responses":{},"parameters":{},"securityDefinitions":{"Bearer":{"name":"Authorization","in":"header","type":"apiKey","description":"Please prefix your JWT token with the word Bearer (and a space)."}},"tags":[{"name":"Authentication","description":"Authentication endpoints for user management"},{"name":"Health","description":"Health check endpoint"},{"name":"Sample","description":"Sample endpoint with full crud"}],"security":[{"Bearer":[]}]}